fix(meta): mixin dependency not installed
Ideally, both npm and yarn would offer an automatic way to install peer dependencies, but sadly this isn't the case. Developers require additional "plugins" that handles this, which is so strange when comparing this to PHP's composer!

Therefore, sadly, the mixins pgk is now a regular dependency. All future packages will also suffer the same treatment.